move class - CsvFormatter
