+++ src/main/java/co/fineants/api/global/common/csv/CsvFileConfig.java
@@ -7,7 +7,6 @@ import org.springframework.context.annotation.Configuration;
 
 import co.fineants.api.domain.dividend.domain.entity.StockDividend;
 import co.fineants.api.domain.stock.domain.entity.Stock;
-import co.fineants.api.infra.s3.service.imple.CsvFormatter;
 
 @Configuration
 public class CsvFileConfig {
