+++ src/main/java/co/fineants/api/infra/s3/service/imple/AmazonS3WriteStockService.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
 import co.fineants.api.domain.stock.domain.entity.Stock;
+import co.fineants.api.global.common.csv.CsvFormatter;
 import co.fineants.api.infra.s3.service.RemoteFileUploader;
 import co.fineants.api.infra.s3.service.WriteStockService;
