+++ src/test/java/co/fineants/api/infra/s3/service/imple/AmazonS3WriteDividendServiceTest.java
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 
 import co.fineants.api.domain.dividend.domain.entity.StockDividend;
+import co.fineants.api.global.common.csv.CsvFormatter;
 import co.fineants.api.infra.s3.service.RemoteFileUploader;
 import co.fineants.api.infra.s3.service.WriteDividendService;
